Follow pagination links when retrieving results
Results from the GitHub API are paginated (30 results per page).
Whenever the results span over multiple page, a header which specifies
the address of the next batch of results is present in the response.

The HTTPClient now follows this link and fetches up to 5 pages of
results.

If more results are present, a link to GitHub is displayed.